[CHORE] seperate tagColor enum file (#18)
