+++ Hatnya/Screens/StudyRoom/UIComponent/HomeworkListCell.swift
@@ -47,6 +47,8 @@ class HomeworkListCell: UICollectionViewCell {
     
 }
 
+// MARK: - UI method
+
 extension HomeworkListCell {
     
     @objc
@@ -109,36 +111,6 @@ struct HomeworkListCellPreview: PreviewProvider {
     
     static var previews: some View {
         HomeworkListCell().toPreview()
-    }
-
-}
-
-enum TagColor: Int, CaseIterable {
-    case pink
-    case yellow
-    case green
-    case lightBlue
-    case purple
-    
-    static func color(_ tagColor: TagColor) -> UIColor {
-        switch tagColor {
-        case .pink:
-            return .tagPink
-        case .yellow:
-            return .tagYellow
-        case .green:
-            return .tagGreen
-        case .lightBlue:
-            return .tagLightBlue
-        case .purple:
-            return .tagPurple
-        }
-    }
-    
-    static func order(index: Int) -> UIColor {
-        let colorNum = index % Self.allCases.count
-        let orderColor = TagColor(rawValue: colorNum) ?? .pink
-        return Self.color(orderColor)
     }
 
 }
